+++ test/data/test_dataset.py
@@ -89,7 +89,7 @@ def test_converter():
     dataset = FakeDataset(
         num_rows=10,
         stypes=[stype.categorical, stype.numerical,
-                stype.multicategorical]).materialize()
+                stype.multi_categorical]).materialize()
     convert_to_tensor_frame = DataFrameToTensorFrameConverter(
         col_to_stype=dataset.col_to_stype,
         col_stats=dataset.col_stats,
@@ -100,11 +100,16 @@ def test_converter():
     assert len(tf) == len(dataset)
 
 
-def test_multicategorical_materialization():
+def test_multi_categorical_materialization():
     data = {'a': ['A,B', 'B,C,A', '', 'B', None]}
     df = pd.DataFrame(data)
-    dataset = Dataset(df, {'a': stype.multicategorical})
+    dataset = Dataset(df, {'a': stype.multi_categorical})
     dataset.materialize()
+    feat = dataset.tensor_frame.feat_dict[stype.multi_categorical]
+    assert torch.equal(feat[0, 0], torch.tensor([1, 0], device=feat.device))
+    assert torch.equal(feat[1, 0], torch.tensor([0, 2, 1], device=feat.device))
+    assert torch.equal(feat[2, 0], torch.Tensor(device=feat.device))
+    assert torch.equal(feat[4, 0], torch.tensor([-1], device=feat.device))
     assert StatType.COUNT in dataset.col_stats['a']
     assert dataset.col_stats['a'][StatType.COUNT][0] == ['B', 'A', 'C']
     assert dataset.col_stats['a'][StatType.COUNT][1] == [3, 2, 1]