+++ torch_frame/stype.py
@@ -23,7 +23,7 @@ class stype(Enum):
     numerical = 'numerical'
     categorical = 'categorical'
     text_embedded = 'text_embedded'
-    multicategorical = 'multicategorical'
+    multi_categorical = 'multi_categorical'
 
     @property
     def is_text_stype(self) -> bool:
@@ -34,10 +34,10 @@ class stype(Enum):
         r""" This property indicates if the data of an stype is stored in
         :class:`torch_frame.data.MultiNestedTensor`.
         """
-        return self in [stype.multicategorical]
+        return self in [stype.multi_categorical]
 
 
 numerical = stype('numerical')
 categorical = stype('categorical')
 text_embedded = stype('text_embedded')
-multicategorical = stype('multicategorical')
+multi_categorical = stype('multi_categorical')